refactor: 회원관리 정보 폴더 구조 정리
