+++ src/pages/memberManagement/MemberInfoPage.tsx
@@ -1,7 +1,7 @@
 import Title from '@/components/Common/Title';
-import ContractInfoBar from '@/components/MemberInfo/ContractInfoBar';
 import MemberInfoBasicTable from '@/components/MemberInfo/MemberBasicInfoTable';
-import MemberInfoButtonBar from '@/components/MemberInfo/MemberInfoButtonBar';
+import ContractInfoBar from '@/components/MemberInfo/MemberInfoCommon/ContractInfoBar';
+import MemberInfoButtonBar from '@/components/MemberInfo/MemberInfoCommon/MemberInfoButtonBar';
 import WorkContractTable from '@/components/MemberInfo/WorkContractTable';
 
 export default function MemberInfoPage() {
